requirements.txt
pandas
jupyter
openpyxl
prettytable
plotly
pycaret
pyarrow
numpy
plotly
matplotlib
chardet
meteostat
awswrangler
stldecompose
prophet
scikit-learn
seaborn
pycaret[full]
pyarrow==15.0.0
streamlit-pandas-profiling
ydata-profiling
pydantic-settings

